chore(gui): aligned group creation explanation w/ updated design
