+++ frontend/src/js/components/devices/group-management/CreateGroupExplainer.test.tsx
@@ -19,7 +19,7 @@ import CreateGroupExplainer from './CreateGroupExplainer';
 
 describe('CreateGroupExplainer Component', () => {
   it('renders correctly', async () => {
-    const { baseElement } = render(<CreateGroupExplainer isEnterprise={false} onClose={vi.fn} />);
+    const { baseElement } = render(<CreateGroupExplainer onClose={vi.fn} />);
     const view = baseElement.getElementsByClassName('MuiDialog-root')[0];
     expect(view).toMatchSnapshot();
     expect(view).toEqual(expect.not.stringMatching(undefineds));
