--- frontend/src/js/components/devices/group-management/CreateGroupExplainerContent.test.tsx
@@ -1,27 +1,0 @@
-// Copyright 2020 Northern.tech AS
-//
-//    Licensed under the Apache License, Version 2.0 (the "License");
-//    you may not use this file except in compliance with the License.
-//    You may obtain a copy of the License at
-//
-//        http://www.apache.org/licenses/LICENSE-2.0
-//
-//    Unless required by applicable law or agreed to in writing, software
-//    distributed under the License is distributed on an "AS IS" BASIS,
-//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//    See the License for the specific language governing permissions and
-//    limitations under the License.
-import { render } from '@/testUtils';
-import { undefineds } from '@northern.tech/testing/mockData';
-import { vi } from 'vitest';
-
-import CreateGroupExplainerContent from './CreateGroupExplainerContent';
-
-describe('CreateGroupExplainerContent Component', () => {
-  it('renders correctly', async () => {
-    const { baseElement } = render(<CreateGroupExplainerContent isEnterprise={false} onClose={vi.fn} />);
-    const view = baseElement.getElementsByClassName('MuiDialog-root')[0];
-    expect(view).toMatchSnapshot();
-    expect(view).toEqual(expect.not.stringMatching(undefineds));
-  });
-});
